pyproject.toml
[tool.poetry]
name = "drf-operation-log"
version = "0.1.0"
description = "Operation log for drf serializers."
authors = ["aiden_lu <aiden_lu@wochacha.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
Django = "^3.2"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

